1. 핵사고날 아키텍처에 적절한 패키지 할당
